move varbin_datum and varbin_to_arrow to separate module
